Fix two locale code matching / 2 char locales
A bug introduced in 1c7b373 set a 2 character locale
when the browser requested a 4 character locale that had
translations available. This resulted in translations failing
to be displayed. This change makes us only fall back to trying
to match the first two characters of the locale if the 4 char
locale was not found. Furthermore, it sets the locale to whatever
the application calls that locale, and not the two character
locale which may not have a translation available.