Remove hookItemsBrowseSql
This is no longer necessary now that the ExperimentalBeijing plugin
adds translations to the search texts and can use those to do searches
for items/browse as well. This will reduce the maintenance burden of the
fork, as all it needs to handle is getting the lang parameter from $_GET
and properly get the default locale. We may want to see if we can at
least get the latter merged upstream.